Moving some tests to outgoing_connection_test.go
