+++ services/gossip/adapter/tcp/direct_transport_test.go
@@ -95,6 +95,47 @@ func TestDirectTransport_SupportsTopologyChangeInRuntime(t *testing.T) {
 	})
 }
 
+func TestDirectTransport_SupportsBroadcastTransmissions(t *testing.T) {
+	with.Concurrency(t, func(ctx context.Context, harness *with.ConcurrencyHarness) {
+		node1 := aNode(ctx, harness.Logger)
+		node2 := aNode(ctx, harness.Logger)
+		node3 := aNode(ctx, harness.Logger)
+		superviseAll(harness, node1, node2, node3)
+		defer shutdownAll(ctx, node1, node2, node3)
+
+		waitForAllNodesToSatisfy(t, "server did not start", func(node *nodeHarness) bool { return node.transport.IsServerListening() }, node1, node2, node3)
+
+		firstTopology := aTopologyContaining(node1, node2, node3)
+		node1.transport.UpdateTopology(ctx, firstTopology)
+		node2.transport.UpdateTopology(ctx, firstTopology)
+		node3.transport.UpdateTopology(ctx, firstTopology)
+
+		waitForAllNodesToSatisfy(t,
+			"expected all nodes to have peers added",
+			func(node *nodeHarness) bool { return len(node.transport.outgoingConnections.activeConnections) > 0 },
+			node1, node2, node3)
+
+		waitForAllNodesToSatisfy(t,
+			"expected all outgoing queues to become enabled after topology change",
+			func(node *nodeHarness) bool { return node.transport.allOutgoingQueuesEnabled() },
+			node1, node2, node3)
+
+		payloads := aMessage()
+
+		node1.listener.ExpectNotReceive()
+		node2.listener.ExpectReceive(payloads)
+		node3.listener.ExpectReceive(payloads)
+		require.NoError(t, node1.transport.Send(ctx, &adapter.TransportData{
+			SenderNodeAddress: node1.address,
+			RecipientMode:     gossipmessages.RECIPIENT_LIST_MODE_BROADCAST,
+			Payloads:          payloads,
+		}))
+
+		require.NoError(t, test.ConsistentlyVerify(test.EVENTUALLY_ADAPTER_TIMEOUT, node1.listener), "message was sent to self node")
+		require.NoError(t, test.EventuallyVerify(test.EVENTUALLY_ADAPTER_TIMEOUT, node2.listener, node3.listener), "message was not sent to target node")
+	})
+}
+
 type nodeHarness struct {
 	transport      *DirectTransport
 	address        primitives.NodeAddress
